Added a script magic-diff to compute pairwise diff
Why? The API server provides `search` capability, and the search results
are limited to similarity values greater than (or equal to) 0.7.

Sometimes a user may simply want to compare two specific files.

This is achieved by adding a script `magic_diff.py` that takes a list of
hashes and computes pairwise diff between all of them. Since similarity
measure is symmetric, it computes only the upper diagonal values.

This is the very first commit in this project, so it does a bit more
than just this one script. It has the capability to cache data
returned from MAGIC server. This is helpful when a user may operate on
the same files multiple times.

This is achieved by using a wrapper around `cythereal-magic` package.
At this time we have capability to query and cache only binary genomics.