Added ability to choose type of hash to show in the output
Why? Not everyone uses sha1 to track files. Some use md5 and others use
sha256. Though MAGIC uses sha1, the ability to choose what hash to use
should be helpful to users.

This was achieved by making the list file a CSV file. The file must
contain a column for 'sha1'. And it may contain other hashes.

The hash to be used is selectable via a command line parameter
`--display-hash`.

Modified example-hashes.txt to make it consistent.